refactor(providers): restructure source layout for filesystem clarity
- Consolidate definitions/ directory into single provider-definitions.ts
- Extract shared adapter infrastructure (output, text-adapter,
  chat-completions-adapter, adapter-types) from adapters/ to shared/
- Promote adapter-registry from adapters/registry.ts to package-root
- Inline provider definitions into concrete provider modules
- Update all import paths across provider and adapter files
- Add shared/index.ts barrel for re-exports
