+++ self/subcortex/providers/src/adapters/anthropic-adapter.ts
@@ -4,14 +4,14 @@
  * ProviderAdapter for the Anthropic Messages API.
  */
 import type { ILogChannel, TraceId } from '@nous/shared';
-import type { ParsedModelOutput } from './output.js';
+import type { ParsedModelOutput } from '../shared/output.js';
 import {
   defineProviderAdapter,
   type AdapterCapabilities,
   type AdapterFormatInput,
   type AdapterFormattedRequest,
   type ProviderAdapter,
-} from './types.js';
+} from '../shared/adapter-types.js';
 
 const ANTHROPIC_CAPABILITIES: AdapterCapabilities = {
   nativeToolUse: true,
