+++ self/subcortex/providers/src/index.ts
@@ -3,7 +3,9 @@
  */
 export { AnthropicProvider } from './anthropic-provider.js';
 export * from './adapters/index.js';
-export * from './definitions/index.js';
+export * from './adapter-registry.js';
+export * from './provider-definitions.js';
+export * from './shared/index.js';
 export { OllamaProvider } from './ollama-provider.js';
 export { ChatCompletionsProvider } from './chat-completions-provider.js';
 export { ProviderRegistry } from './provider-registry.js';
