+++ self/subcortex/providers/src/shared/output.ts
