+++ self/subcortex/providers/src/shared/text-adapter.ts
@@ -6,7 +6,7 @@ import {
   type AdapterFormatInput,
   type AdapterFormattedRequest,
   type ProviderAdapter,
-} from './types.js';
+} from './adapter-types.js';
 
 const TEXT_ADAPTER_CAPABILITIES: AdapterCapabilities = {
   nativeToolUse: false,
